+++ src/Mandrill.php
@@ -80,7 +80,7 @@ class Mandrill implements EmailSenderInterface
                 ($result['status'] === 'sent' || $result['status'] === 'queued' || $result['status'] === 'scheduled')
             ) {
                 if ($this->logger) {
-                    $this->logger->info("Email sent: '{$email->getSubject()}'", $email);
+                    $this->logger->info("Email sent: '{$email->getSubject()}'", $email->toArray());
                 }
             } else {
                 if (!$result || !$result['status']) {
@@ -98,12 +98,12 @@ class Mandrill implements EmailSenderInterface
             }
         } catch (Exception $e) {
             if ($this->logger) {
-                $this->logger->info("Email error: '{$e->getMessage()}'", $email);
+                $this->logger->info("Email error: '{$e->getMessage()}'", $email->toArray());
             }
             throw $e;
         } catch (\Mandrill_Error $e) {
             if ($this->logger) {
-                $this->logger->info("Email error: '{$e->getMessage()}'", $email);
+                $this->logger->info("Email error: '{$e->getMessage()}'", $email->toArray());
             }
             throw new InvalidRequestException($e->getMessage(), 601, $e);
         }